% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/th_test.R
\name{generate_example_series}
\alias{generate_example_series}
\title{Generate forward-model time series across parameter combinations}
\usage{
generate_example_series(
  aquifer = generate_example_aquifer(),
  scenarios = list(diel = list(k = c(10)/86400, dispersivity = c(0.001, 0.01, 0.1),
    headGrad = 0.1, mean = 22, amplitude = 6, phase = 86400 + 3600, period = 86400, xVals
    = c(0, 0.33, 0.66, 1), tVals = seq(0, 3600 * 48, 3600)), annual = list(k =
    c(100)/86400, dispersivity = c(100, 1000), headGrad = 0.01, mean = 11.78, amplitude =
    8.87, phase = 1702226, period = 86400 * 365, xVals = c(0, 120, 240, 360), tVals =
    seq(0, 2 * 365 * 86400, 86400)))
)
}
\arguments{
\item{aquifer}{A thAquifer object. Default uses generate_example_aquifer().}

\item{scenarios}{A named list of scenario specifications.}
}
\value{
A named list of lists. Top level is scenario name (e.g., 'diel',
  'annual'). Each contains one thSeries per parameter combination.
}
\description{
Creates thSeries objects for factorial combinations of hydraulic
conductivity and dispersivity, at both diel and annual time scales.
}
