% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3Objects.R
\name{thAquifer}
\alias{thAquifer}
\title{Create an aquifer properties object}
\usage{
thAquifer(
  porosity,
  thermCond_sed,
  thermCond_h2o,
  spHeat_sed,
  spHeat_h2o,
  density_sed,
  density_h2o,
  specificUnits = thUnits()
)
}
\arguments{
\item{porosity}{Porosity [L3 L-3] (volume fraction)}

\item{thermCond_sed}{Thermal conductivity of sediment [E t-1 L-1 T-1]}

\item{thermCond_h2o}{Thermal conductivity of water [E t-1 L-1 T-1]}

\item{spHeat_sed}{Specific heat of sediment [E M-1 T-1]}

\item{spHeat_h2o}{Specific heat of water [E M-1 T-1]}

\item{density_sed}{Density of sediment [M L-3]}

\item{density_h2o}{Density of water [M L-3]}

\item{specificUnits}{A thUnits object for labeling}
}
\value{
A thAquifer S3 object
}
\description{
Stores user-specified material properties and computes bulk properties
(volumetric heat capacities, bulk thermal conductivity) as porosity-weighted
averages of sediment and water values.
}
