% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3Objects.R
\name{thHydro}
\alias{thHydro}
\title{Create a hydrology/transport properties object}
\usage{
thHydro(
  hydCond,
  dispersivity = 0.001,
  headGrad,
  aquifer,
  specificUnits = thUnits()
)
}
\arguments{
\item{hydCond}{Hydraulic conductivity [L t-1]}

\item{dispersivity}{Thermal dispersivity [L] (default 0.001)}

\item{headGrad}{Hydraulic head gradient [L L-1]}

\item{aquifer}{A thAquifer object}

\item{specificUnits}{A thUnits object}
}
\value{
A thHydro S3 object
}
\description{
Combines user-specified flow parameters with aquifer material properties
to derive Darcy flux, water velocity, advective thermal velocity, and
thermal diffusivity components (conductive, dispersive, effective).
}
