% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/th_test.R
\name{generate_example_aquifer}
\alias{generate_example_aquifer}
\title{Create an example thAquifer with basalt-like defaults}
\usage{
generate_example_aquifer(
  porosity = 0.25,
  thermCond_sed = 0.0016,
  thermCond_h2o = 0.000598,
  spHeat_sed = 0.84,
  spHeat_h2o = 4.186,
  density_sed = 3000,
  density_h2o = 998
)
}
\value{
A thAquifer object.
}
\description{
Default values from Columbia River basalt (Burns et al. 2015, GeoFluids v.15).
}
\references{
Burns, E. R., Williams, C. F., Ingebritsen, S. E., Voss, C. I., Spane,
F. A., & DeAngelo, J. (2015). Understanding heat and groundwater flow
through continental flood basalt provinces: insights gained from alternative
models of permeability/depth relationships for the Columbia Plateau, USA.
Geofluids, 15, 120-138.
}
