% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3Objects.R
\name{thAquifer}
\alias{thAquifer}
\alias{thBoundary}
\alias{thUnits}
\title{Temperheic Aquifer, Boundary, and Units objects}
\usage{
thAquifer(porosity, thermCond_sed, thermCond_h2o, spHeat_sed, spHeat_h2o,
  density_sed, density_h2o, specificUnits = thUnits())

thBoundary(mean, amplitude, phase, period, specificUnits = thUnits())

thUnits(L = "m", M = "kg", t = "s", T = "degC", E = "kJ")
}
\arguments{
\item{porosity}{Porosity (L3 L-3) of the aquifer; ratio of water volume to
unit aquifer volume under saturated conditions.}

\item{thermCond_sed}{Thermal conductivity (E t-1 L-1 T-1) of the sediment.}

\item{thermCond_h2o}{Thermal conductivity (E t-1 L-1 T-1) of water.}

\item{spHeat_sed}{Specific heat (E M-1 T-1) of the sediment.}

\item{spHeat_h2o}{Specific heat (E M-1 T-1) of water.}

\item{density_sed}{Density (M L-3) of sediment.}

\item{density_h2o}{Density (M L-3) of water.}

\item{specificUnits}{A \code{thUnits} object generated by calling
\code{thUnits()}. Specifies the units of length (L), mass (M), time (t),
temperature (T), and energy (E) used in the numeric arguments.  A
\code{thUnits} object is used only to label the values provided by the user,
in hopes of helping the user avoid errors.  The \code{temperheic} package
makes no conversions.  The user must ensure that all arguments passed to
\code{thAquifer()} and \code{thBoundary()} have units consistent with
\code{specificUnits}.}

\item{L}{A chararcter string label for units of length.  Default is 'm' for
meters.}

\item{M}{A chararcter string label for units of mass.  Default is 'kg' for
kilograms.}

\item{t}{A chararcter string label for units of time.  Default is 's' for
seconds.}

\item{T}{A chararcter string label for units of temperature.  Default is
"degC" for degrees celsius.}

\item{E}{A chararcter string label for units of Energy. Default is "kJ" for
kilojoules.}

\item{hydCond}{Hydraulic conductivity (L t-1) (either vertical or horizontal)
of the aquifer.}

\item{dispersivity}{Dispersivity (L) of the aquifer, appropriate for the the
scale of the flow path observed (see Gelher 20XX).}

\item{headGrad}{Average head gradient (L L-1) in the same dimension (vertical
or horizonal) for the aquifer.}
}
\value{
\code{thAquifer()} returns a S3 object of class \code{thAquifer} which
 includes all specified and derived parameters for the aquifer

\code{thBoundary} returns an S3 object of class \code{thBoundary}, which
  includes the specified and derived parameters describing a temperature
  signal at the aquifer boundary.

\code{thUnits()} return an S3 object of class \code{thUnits}, which
  includes specific units of length, mass, time, temperature, and energy. The
  values in a \code{thUnits} object are used to create unit labels for
  \code{thAquifer} and \code{thBoundary} parameter values.
}
\description{
The \code{temperheic} package relies on three different S3
 objects: \code{thAquifer}, \code{thBoundary}, and \code{thUnits}.  The "th"
 prefix is short for "temperheic".  A \code{thAquifer} object contains
 user-specified and derived aquifer properties.  A \code{thBoundary} object
 contains user-specified and derived values that describe a time varying
 temperature boudary as a cosine function.  A \code{thUnits} object describes
 the units of measure used to specify \code{thAquifer} and \code{thBoundary}
 objects.

 NOTE THAT THE USER IS RESPONSIBLE FOR ENSURING THAT THE UNITS OF ALL
 ARGUMENTS PASSED TO \code{TEMPERHEIC} PACKAGE FUNCTIONS ARE CONSISTENT. See
 Details (below) for more information on units.
}
\details{
\code{thAquifer} objects include values for the arguments provided by
 the user (see 'Arguments', above), along with the following derived values:

 \itemize{

 \item{\code{darcyFlux}}{  Darcy Flux (L t-1) is flux rate of water through
 the aquifer along the hydrologic gradient. Calculated as:
 \code{hydCond*headGrad}.}

 \item{\code{density_bulk}}{  Density of the bulk medium (M L-3) is the
 average density of water and sediment, weighted by porosity.  Calculated as:
 \code{density_sed*(1-porosity)+density_h2o*porosity}.}

 \item{\code{spHeat_bulk}}{  Specific heat of the bulk medium (E M-1 T-1) is
 the average specific heat of water and sediments, weighted by porosity.
 Calculated as: \code{spHeat_sed*(1-porosity)+spHeat_h2o*porosity}.}

 \item{\code{velocity_h2o}}{  Water velocity (L T-1) is the aerial averaged
 rate of water movement along the hydrologic gradient in the aquifer.
 Calculated as: \code{darcyFlux/porosity}.}

 \item{\code{darcyFlux_heat}}{  Heat velocity (L T-1) is the aerially
 averaged rate of heat movement along the hydraulic gradient in the aquifer.
 Calculated as: \code{darcyFlux_h2o*(volHeatCap_h2o/volHeatCap_bulk)} Note
 that the temperheic package represents 1D movement of water and heat, so
 that heat gradients and hydrologic gradients are always aligned.}

 \item{\code{volHeatCap_h2o, volHeatCap_sed}}{  The volumetric heat capacity
 of of water, and of sediment [E L-3 T-1] are calcuated from the
 corresponding values for specific heat and density:
 \code{spHeat_<x>*density_<x>}.}

 \item{\code{volHeatCap_bulk} {  The volumentric heat capacity of the bulk
 medium [E L-3 T-1] is calcuated as the porosity-weighted average of
 volumetric heat capacity of water and sediment:
 volHeatCap_sed*(1-porosity)+volHeatCap_h2o*porosity}.}

 }

\code{thBoundary} objects include values for the arguments provided
  by the user (see 'Arguments', above), along with the following derived
  value:

  \itemize{

  \item{\code{frequency}}{  Frequency (t-1) of the temperature signal at the
  boundary.  Calculated as: \code{1/period}.}

  }

\code{thUnits} objects include unit labels for the parameter values
  stored in \code{thAquifer} and \code{thBoundary} objects.  Labels for units
  of length (L), mass (M), time (t), temperature (T), and Energy (E) are
  specified when creating a \code{thUnits} object.  The \code{thUnits} object
  can be stored in a variable and subsequently passed to \code{thAquifer()}
  and \code{thBoundary()}.  Note that the \code{thUnits} object is used ONLY
  to create lables for \code{thAquifer} and \code{thBoundary} values.  No
  attempt is made to convert to common units.  The user is responsible for
  ensuring that the units of all values in \code{thAquifer} and
  \code{thBoundary} objects agree with the unit labels displayed when the
  object is printed.
}
