% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thSeries.R
\name{thObservedSeries}
\alias{thObservedSeries}
\alias{thSeries}
\title{Create thSeries and thObservation Objects.}
\usage{
thObservedSeries(empiricalData, xVals, period, aquifer, hydro, nmin,
  freq = (2 * pi)/period, optimizeRange = c(-period/8, period * 7/8),
  specificUnits = thUnits())

thSeries(signal, xVals, tVals, specificUnits)
}
\description{
thObservations are \code{\link{zoo}} objects.  Each column contains a
temperature record for a different well in the aquifer.  thSeries
}
